TRAD-143 rename pkg directory
