+++ benchmark/internal/transaction/submit.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"github.com/bloXroute-Labs/solana-trader-client-go/benchmark/internal/logger"
 	"github.com/bloXroute-Labs/solana-trader-client-go/benchmark/internal/utils"
-	"github.com/bloXroute-Labs/solana-trader-client-go/pkg/provider"
-	"github.com/bloXroute-Labs/solana-trader-client-go/pkg/transaction"
 	pb "github.com/bloXroute-Labs/solana-trader-client-go/proto"
+	"github.com/bloXroute-Labs/solana-trader-client-go/provider"
+	"github.com/bloXroute-Labs/solana-trader-client-go/transaction"
 	"github.com/gagliardetto/solana-go"
 	solanarpc "github.com/gagliardetto/solana-go/rpc"
 	"strconv"
