+++ benchmark/txcompare/main.go
@@ -7,7 +7,7 @@ import (
 	"github.com/bloXroute-Labs/solana-trader-client-go/benchmark/internal/logger"
 	"github.com/bloXroute-Labs/solana-trader-client-go/benchmark/internal/transaction"
 	"github.com/bloXroute-Labs/solana-trader-client-go/benchmark/internal/utils"
-	"github.com/bloXroute-Labs/solana-trader-client-go/pkg/provider"
+	"github.com/bloXroute-Labs/solana-trader-client-go/provider"
 	"github.com/gagliardetto/solana-go"
 	"github.com/urfave/cli/v2"
 	"os"
