+++ connections/common.go
