+++ provider_test/grpc_test.go
@@ -1,44 +1,40 @@
 package provider
 
 import (
-	"github.com/stretchr/testify/require"
+	"github.com/bloXroute-Labs/solana-trader-client-go/provider"
 	"os"
 	"testing"
 	"time"
 
-	"github.com/bloXroute-Labs/solana-trader-client-go/pkg/provider"
 	"github.com/gagliardetto/solana-go"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestWS_New(t *testing.T) {
+func TestGRPC_New(t *testing.T) {
 	pk, err := solana.NewRandomPrivateKey()
 	assert.NotNil(t, pk)
 	assert.Nil(t, err)
 
-	err = os.Setenv("PRIVATE_KEY", pk.String())
-	require.Nil(t, err)
+	os.Setenv("PRIVATE_KEY", pk.String())
 
-	c, err := provider.NewWSClient()
+	c, err := provider.NewGRPCClient()
 	assert.NotNil(t, c)
 	assert.Nil(t, err)
 
-	err = os.Unsetenv("PRIVATE_KEY")
-	require.Nil(t, err)
+	os.Unsetenv("PRIVATE_KEY")
 }
 
-func TestWS_NewWithOpts(t *testing.T) {
+func TestGRPC_NewWithOpts(t *testing.T) {
 	pk, err := solana.NewRandomPrivateKey()
 	assert.NotNil(t, pk)
 	assert.Nil(t, err)
 
 	o := provider.RPCOpts{
-		Endpoint:   provider.MainnetWS,
+		Endpoint:   provider.MainnetGRPC,
 		Timeout:    time.Second,
 		PrivateKey: &pk,
 	}
-
-	c, err := provider.NewWSClientWithOpts(o)
+	c, err := provider.NewGRPCClientWithOpts(o)
 	assert.NotNil(t, c)
 	assert.Nil(t, err)
 }
