+++ provider_test/ws_test.go
@@ -1,40 +1,44 @@
 package provider
 
 import (
+	"github.com/bloXroute-Labs/solana-trader-client-go/provider"
+	"github.com/stretchr/testify/require"
 	"os"
 	"testing"
 	"time"
 
-	"github.com/bloXroute-Labs/solana-trader-client-go/pkg/provider"
 	"github.com/gagliardetto/solana-go"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestHTTP_New(t *testing.T) {
+func TestWS_New(t *testing.T) {
 	pk, err := solana.NewRandomPrivateKey()
 	assert.NotNil(t, pk)
 	assert.Nil(t, err)
 
-	os.Setenv("PRIVATE_KEY", pk.String())
+	err = os.Setenv("PRIVATE_KEY", pk.String())
+	require.Nil(t, err)
 
-	c := provider.NewHTTPClient()
+	c, err := provider.NewWSClient()
 	assert.NotNil(t, c)
 	assert.Nil(t, err)
 
-	os.Unsetenv("PRIVATE_KEY")
+	err = os.Unsetenv("PRIVATE_KEY")
+	require.Nil(t, err)
 }
 
-func TestHTTP_NewWithOpts(t *testing.T) {
+func TestWS_NewWithOpts(t *testing.T) {
 	pk, err := solana.NewRandomPrivateKey()
 	assert.NotNil(t, pk)
 	assert.Nil(t, err)
 
 	o := provider.RPCOpts{
-		Endpoint:   provider.MainnetHTTP,
+		Endpoint:   provider.MainnetWS,
 		Timeout:    time.Second,
 		PrivateKey: &pk,
 	}
 
-	c := provider.NewHTTPClientWithOpts(nil, o)
+	c, err := provider.NewWSClientWithOpts(o)
 	assert.NotNil(t, c)
+	assert.Nil(t, err)
 }
